SH-160 Refactor project structure and CMakeLists.txt
Changed project structure for better separation of SmartHome components. Implemented CMakeLists.txt for subdirectories and updated the main CMakeLists.txt accordingly.
