+++ central_unit/CMakeLists.txt
@@ -1,16 +1,37 @@
 cmake_minimum_required(VERSION 3.21)
+project(SmartHome)
+
+set(CMAKE_CXX_STANDARD 23)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF)
+
 
 # Policy for boost compatibility
-if(POLICY CMP0167)
+if (POLICY CMP0167)
     cmake_policy(SET CMP0167 OLD)
-endif()
+endif ()
 
-project(SmartHome)
 
 # Default value for TARGET_DEVICE
 set(TARGET_DEVICE "NATIVE" CACHE STRING "Target device type")
 # Possible values for TARGET_DEVICE
 set_property(CACHE TARGET_DEVICE PROPERTY STRINGS "NATIVE" "RPI_ZERO_2W")
+
+# Build options
+option(BUILD_CORE "Build SmartHome daemon" ON)
+option(BUILD_MEDIATOR "Build module mediator" ON)
+
+# Boost config
+set(Boost_NO_WARN_NEW_VERSIONS ON)
+set(Boost_USE_MULTITHREADED ON)
+set(Boost_USE_STATIC_RUNTIME OFF)
+
+# Static boost linking for release
+if (CMAKE_BUILD_TYPE STREQUAL "Release")
+    set(Boost_USE_STATIC_LIBS ON)
+else ()
+    set(Boost_USE_STATIC_LIBS OFF)
+endif ()
 
 # Config info
 message(STATUS "==================[ Build Configuration ]=====================")
@@ -22,98 +43,49 @@ message(STATUS "CMAKE_CXX_COMPILER: ${CMAKE_CXX_COMPILER}")
 message(STATUS "Target Device: ${TARGET_DEVICE}")
 message(STATUS "==============================================================")
 
-set(CMAKE_CXX_STANDARD 23)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_CXX_EXTENSIONS OFF)
 
-# Boost config
-set(Boost_NO_WARN_NEW_VERSIONS ON)
-set(Boost_USE_MULTITHREADED ON)
-set(Boost_USE_STATIC_RUNTIME OFF)
+add_subdirectory(src/common)
 
-# Static boost linking for release
-if(CMAKE_BUILD_TYPE STREQUAL "Release")
-    set(Boost_USE_STATIC_LIBS ON)
-else()
-    set(Boost_USE_STATIC_LIBS OFF)
-endif()
+if (BUILD_CORE)
+    add_subdirectory(src/core)
+endif ()
 
-# Boost required
-find_package(Boost 1.74 REQUIRED COMPONENTS
-        system
-        thread
-        filesystem
-        program_options
-        chrono
-)
-
-# Yaml-cpp required
-find_package(yaml-cpp REQUIRED)
-
-add_executable(SmartHome
-        src/main.cpp
-        src/smart_home_core.cpp
-        src/ipc/smart_home_tcp_server.cpp
-        src/ipc/smart_home_tcp_connection.cpp
-        src/smart_home_config.cpp
-)
-add_executable(ModuleMediator
-        src/module_mediator.cpp
-        src/smart_home_config.cpp
-)
-
-
-target_link_libraries(SmartHome PRIVATE
-        Boost::system
-        Boost::thread
-        Boost::filesystem
-        Boost::program_options
-        Boost::chrono
-
-        yaml-cpp
-
-        pthread
-)
-
-target_link_libraries(ModuleMediator PRIVATE
-        Boost::system
-        Boost::program_options
-        pthread
-)
-
-
-# Platform based config
-if(CMAKE_SYSTEM_PROCESSOR STREQUAL "aarch64")
-    target_compile_options(SmartHome PRIVATE
-            -march=armv8-a
-    )
-endif()
-
-if(TARGET_DEVICE STREQUAL "RPI_ZERO_2W")
-    target_compile_options(SmartHome PRIVATE
-            -mtune=cortex-a53
-    )
+if (BUILD_MEDIATOR)
+    add_subdirectory(src/mediator)
 endif ()
 
 
-if(CMAKE_BUILD_TYPE STREQUAL "Debug")
+# Platform based compile config
+if (CMAKE_SYSTEM_PROCESSOR STREQUAL "aarch64")
+    add_compile_options(-march=armv8-a)
+endif ()
+
+if (TARGET_DEVICE STREQUAL "RPI_ZERO_2W")
+    add_compile_options(-mtune=cortex-a53)
+endif ()
+
+
+if (CMAKE_BUILD_TYPE STREQUAL "Debug")
     # Additional options for debuging
-    target_compile_options(SmartHome PRIVATE
+    add_compile_options(
             -O0
             -g3
             -DDEBUG
     )
 else ()
     # Additional options for release optimization
-    target_compile_options(SmartHome PRIVATE
+    add_compile_options(
             -O2
             -DNDEBUG
             -ffunction-sections
             -fdata-sections
     )
-    target_link_options(SmartHome PRIVATE
-            -Wl,--gc-sections
-    )
+    add_link_options(-Wl,--gc-sections)
     set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -s")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -s")
 endif ()
+
+message(STATUS "======================[ Feature summary ]=======================")
+include(FeatureSummary)
+feature_summary(WHAT ALL)
+message(STATUS "================================================================")
